pyproject.toml
[tool.poetry]
name = "pytorch_lightning_example"
version = "0.1.0"
description = ""
authors = ["chatflip <oo.chat.flip@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8,<3.10"
torch = {version = "1.13.1+cu117", source = "torch_cu117"}
torchvision = {version = "0.14.1+cu117", source = "torch_cu117"}
pytorch-lightning = "1.9.4"
timm = "0.6.12"
segmentation-models-pytorch = "0.3.2"
albumentations = "1.3.0"
mlflow = "2.2.2"
hydra-core = "1.3.2"
opencv-python = "4.7.0.72"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pysen = "^0.10"
black = "21.10b0"
flake8 = "4.0.1"
isort = "5.10.1"
mypy = "0.910"
click = "8.0.2"

[[tool.poetry.source]]
name = "torch_cu117"
url = "https://download.pytorch.org/whl/cu117/"
default = false
secondary = true

[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "very_strict"
line_length = 88
py_version = "py39"
[[tool.pysen.lint.mypy_targets]]
paths = [
    "1_classification_food101/"
]

[tool.pysen.lint.source]
include_globs = [
    "1_classification_food101/src/**/*.py"
]

[tool.isort]
target-version = ["py39"]
profile = "black"
line_length = 88

[tool.mypy]
python_version = 3.9
# following setting is same as pysen
# https://github.com/pfnet/pysen/blob/main/setup.cfg#L12
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_generics = false
disallow_any_unimported = false
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
strict_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
